refactor : JCF Repository 추상 클래스
