+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFChannelRepository.java
@@ -3,34 +3,5 @@ package com.sprint.mission.discodeit.repository.jcf;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 
-import java.util.*;
-
-public class JCFChannelRepository implements ChannelRepository {
-    private final Map<UUID, Channel> data = new HashMap<>();
-
-    @Override
-    public Channel save(Channel channel) {
-        data.put(channel.getId(), channel);
-        return channel;
-    }
-
-    @Override
-    public Channel findById(UUID id) {
-        return data.getOrDefault(id, null);
-    }
-
-    @Override
-    public List<Channel> findAll() {
-        return new ArrayList<>(data.values());
-    }
-
-    @Override
-    public void deleteById(UUID id) {
-        data.remove(id);
-    }
-
-    @Override
-    public void deleteAll() {
-        data.clear();
-    }
+public class JCFChannelRepository extends AbstractJCFRepository<Channel> implements ChannelRepository {
 }
