+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFMessageRepository.java
@@ -3,34 +3,5 @@ package com.sprint.mission.discodeit.repository.jcf;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 
-import java.util.*;
-
-public class JCFMessageRepository implements MessageRepository {
-    private final Map<UUID, Message> data = new HashMap<>();
-
-    @Override
-    public Message save(Message message) {
-        data.put(message.getId(), message);
-        return message;
-    }
-
-    @Override
-    public Message findById(UUID id) {
-        return data.getOrDefault(id, null);
-    }
-
-    @Override
-    public List<Message> findAll() {
-        return new ArrayList<>(data.values());
-    }
-
-    @Override
-    public void deleteById(UUID id) {
-        data.remove(id);
-    }
-
-    @Override
-    public void deleteAll() {
-        data.clear();
-    }
+public class JCFMessageRepository extends AbstractJCFRepository<Message> implements MessageRepository {
 }
