+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFUserRepository.java
@@ -3,34 +3,5 @@ package com.sprint.mission.discodeit.repository.jcf;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 
-import java.util.*;
-
-public class JCFUserRepository implements UserRepository {
-    private final Map<UUID, User> data = new HashMap<>();
-
-    @Override
-    public User save(User user) {
-        data.put(user.getId(), user);
-        return user;
-    }
-
-    @Override
-    public User findById(UUID id) {
-        return data.getOrDefault(id, null);
-    }
-
-    @Override
-    public List<User> findAll() {
-        return new ArrayList<>(data.values());
-    }
-
-    @Override
-    public void deleteById(UUID id) {
-        data.remove(id);
-    }
-
-    @Override
-    public void deleteAll() {
-        data.clear();
-    }
+public class JCFUserRepository extends AbstractJCFRepository<User> implements UserRepository {
 }
